backend/requirements.txt
asgiref==3.3.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
charset-normalizer==2.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
daphne==3.0.2
Django==3.1.3
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.8.0
django-filter==2.4.0
django-oauth-toolkit==1.3.3
djangorestframework==3.12.2
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
hyperlink==21.0.0
idna==3.2
importlib-metadata==2.0.0
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
Markdown==3.3.3
MarkupSafe==2.0.1
numpy==1.21.2
oauthlib==3.1.1
openpyxl==3.0.7
packaging==21.0
pandas==1.2.3
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
requests==2.26.0
ruamel.yaml==0.17.11
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
xlrd==2.0.1
zipp==3.5.0
zope.interface==5.4.0
